+++ do_feeds.php
@@ -9,21 +9,21 @@
  * @since  1.6.0-fork
  */
 
+namespace Lwt\Interface\Do_Feeds;
+
 require_once 'inc/session_utility.php';
 
 $currentlang = validateLang(processDBParam("filterlang", 'currentlanguage', '', 0));
-$currentsort = processDBParam("sort", 'currentrsssort', '2', 1);
 $currentquery = processSessParam("query", "currentrssquery", '', 0);
 $currentquerymode = processSessParam("query_mode", "currentrssquerymode", 'title,desc,text', 0);
 $currentregexmode = getSettingWithDefault("set-regex-mode");
-$currentpage = processSessParam("page", "currentrsspage", '1', 1);
 $currentfeed = processSessParam("selected_feed", "currentrssfeed", '', 0);
 $wh_query = $currentregexmode . 'like ' .  convert_string_to_sqlsyntax(
     ($currentregexmode == '') ? 
     str_replace("*", "%", mb_strtolower($currentquery, 'UTF-8')) : 
     $currentquery
 );
-switch($currentquerymode){
+switch ($currentquerymode) {
 case 'title,desc,text':
     $wh_query=' and (FlTitle ' . $wh_query . ' or FlDescription ' . $wh_query . ' or FlText ' . $wh_query . ')';
     break;
@@ -33,13 +33,14 @@ case 'title':
 }
 pagestart('My ' . getLanguage($currentlang) . ' Feeds', true);
 
-if($currentquery!=='') {
-    if($currentregexmode!=='') {
-        if(@mysqli_query($GLOBALS["DBCONNECTION"], 'select "test" rlike ' . convert_string_to_sqlsyntax($currentquery))===false) {
+if ($currentquery!=='') {
+    if ($currentregexmode!=='') {
+        if (@mysqli_query($GLOBALS["DBCONNECTION"], 'select "test" rlike ' . convert_string_to_sqlsyntax($currentquery))===false) {
             $currentquery='';
             $wh_query = '';
             unset($_SESSION['currentwordquery']);
-            if(isset($_REQUEST['query'])) { echo '<p id="hide3" style="color:red;text-align:center;">+++ Warning: Invalid Search +++</p>'; 
+            if(isset($_REQUEST['query'])) { 
+                echo '<p id="hide3" style="color:red;text-align:center;">+++ Warning: Invalid Search +++</p>'; 
             }
         }
     }
@@ -52,7 +53,10 @@ $edit_text=0;
 
 $doc = null;
 $text_item = null;
-if (isset($_REQUEST['marked_items']) && is_array($_REQUEST['marked_items'])) {
+
+function dummy_function_1(&$edit_text)
+{
+    global $tbpref;
     $marked_items = implode(',', $_REQUEST['marked_items']);
     $res = do_mysqli_query(
         "SELECT * FROM (
@@ -336,7 +340,7 @@ if (isset($_REQUEST['marked_items']) && is_array($_REQUEST['marked_items'])) {
    </script>
         <?php
     }
-    if($edit_text==1) {
+    if ($edit_text==1) {
         echo '</form>';
     }
     ?>
@@ -344,11 +348,16 @@ if (isset($_REQUEST['marked_items']) && is_array($_REQUEST['marked_items'])) {
 $(".hide_message").delay(2500).slideUp(1000);
 </script>
     <?php
+    return $message;
+}
+
+if (isset($_REQUEST['marked_items']) && is_array($_REQUEST['marked_items'])) {
+    $message = dummy_function_1($edit_text);
 }
 
 
 if (isset($_REQUEST['checked_feeds_save'])) {
-    $message=write_rss_to_db($_REQUEST['feed']);
+    $message = write_rss_to_db($_REQUEST['feed']);
     ?>
 <script type="text/javascript">
 $(".hide_message").delay(2500).slideUp(1000);
@@ -373,18 +382,21 @@ $(".hide_message").delay(2500).slideUp(1000);
 }
 echo error_message_with_hide($message, 0);
 
-if (
-    isset($_REQUEST['load_feed']) || isset($_REQUEST['check_autoupdate']) || 
-    (isset($_REQUEST['markaction']) && $_REQUEST['markaction']=='update')) {
-    load_feeds($currentfeed);
-} else if(empty($edit_text)) {
+function dummy_function_2(
+    $currentlang, $wh_query, $currentquery, $currentfeed, 
+    $currentquerymode, $currentregexmode
+    )
+{
+    global $tbpref, $debug;
+    $currentpage = processSessParam("page", "currentrsspage", '1', 1);
+    $currentsort = processDBParam("sort", 'currentrsssort', '2', 1);
     ?>
 
 <div class="flex-spaced">
     <div title="Import of a single text, max. 65,000 bytes long, with optional audio">
-        <a href="<?php echo $_SERVER['PHP_SELF']; ?>?new=1">
-            <img src="icn/plus-button.png">
-            New Text
+        <a href="edit_feeds.php?new_feed=1">
+            <img src="icn/feed--plus.png">
+            New Feed
         </a>
     </div>
     <div>
@@ -594,6 +606,17 @@ $('img.not_found').on('click', function () {
 });
 </script>
     <?php
+}
+
+if (
+    isset($_REQUEST['load_feed']) || isset($_REQUEST['check_autoupdate']) || 
+    (isset($_REQUEST['markaction']) && $_REQUEST['markaction']=='update')) {
+    load_feeds($currentfeed);
+} else if(empty($edit_text)) {
+    dummy_function_2(
+        $currentlang, $wh_query, $currentquery, $currentfeed, $currentquerymode,
+        $currentregexmode
+    );
 }
 
 pageend();